requirements.txt
python-telegram-bot
yt-dlp
requests